Extract BCryptImportKey from Cng.cs to Interop.BCryptImportKey.cs
